requirements.txt
# Core dependencies
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
langchain-chroma==0.1.0

# Vector stores
chromadb==0.4.22

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pyyaml==6.0.1
requests==2.31.0

# Data processing
numpy==1.24.3
pandas==2.0.3

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1
notebook==7.0.6

# HTTP client
httpx==0.26.0
